# Much faster CPU prompt processing (part 3)