--- a/ggml/src/iqk/iqk_gemm_legacy_quants.cpp
+++ b/ggml/src/iqk/iqk_gemm_legacy_quants.cpp
@@ -1727,30 +1727,29 @@ void iqk_convert_q80_q80_r8(int n, const void * vx, size_t bx, void * vy, int nr
     }
 }
 
-void iqk_convert_q40_q80_r8(int n, const void * vx, size_t bx, void * vy, int nrc_x) {
-    static_assert(QK4_0 == QK8_0);
+template <typename Block, typename Dequantizer>
+void iqk_convert_qX_q80_r8(int n, const void * vx, size_t bx, void * vy, int nrc_x) {
     GGML_ASSERT(n%QK4_0 == 0);
     GGML_ASSERT(nrc_x%8 == 0);
 
-    const int nb = n/QK4_0;
+    const int nb = n/QK8_0;
 
     block_q8_0_r8 * y = (block_q8_0_r8 *)vy;
 
-    const block_q4_0 * x8[8];
+    const Block * x8[8];
 
     uint32_t block[8];
 
+    Dequantizer deq;
+
     for (int ix = 0; ix < nrc_x; ix += 8) {
 
-        for (int k = 0; k < 8; ++k) x8[k] = (const block_q4_0 *)((const char *)vx + (ix + k)*bx);
+        for (int k = 0; k < 8; ++k) x8[k] = (const Block *)((const char *)vx + (ix + k)*bx);
 
         for (int i = 0; i < nb; ++i) {
             for (int k = 0; k < 8; ++k) {
                 y[i].d[k] = x8[k][i].d;
-                auto bits = _mm_loadu_si128((const __m128i *)x8[k][i].qs);
-                auto val  = _mm256_and_si256(MM256_SET_M128I(_mm_srli_epi16(bits, 4), bits), _mm256_set1_epi8(0xf));
-                val = _mm256_add_epi8(val, _mm256_set1_epi8(-8));
-                _mm256_storeu_si256((__m256i *)block, val);
+                _mm256_storeu_si256((__m256i *)block, deq.dequant(x8[k] + i));
                 auto qs = (uint32_t *)y[i].qs;
                 for (int l = 0; l < 4; ++l) {
                     qs[8*l + k +  0] = block[l + 0];
@@ -1787,7 +1786,9 @@ template <typename Dequantizer> void set_functions(std::array<mul_mat_t, IQK_MAX
 
 bool iqk_convert_legacy_quants_q8_r8(int type, int n, const void * vx, size_t bx, void * vy, int nrc_x) {
     switch (type) {
-        case GGML_TYPE_Q4_0: iqk_convert_q40_q80_r8(n, vx, bx, vy, nrc_x); break;
+        case GGML_TYPE_Q4_0: iqk_convert_qX_q80_r8<block_q4_0, Q4_0_Dequantizer>(n, vx, bx, vy, nrc_x); break;
+        case GGML_TYPE_Q5_0: iqk_convert_qX_q80_r8<block_q5_0, Q5_0_Dequantizer>(n, vx, bx, vy, nrc_x); break;
+        //case GGML_TYPE_Q6_0: iqk_convert_qX_q80_r8<block_q6_0, Q6_0_Dequantizer>(n, vx, bx, vy, nrc_x); break;
         case GGML_TYPE_Q8_0: iqk_convert_q80_q80_r8(n, vx, bx, vy, nrc_x); break;
         default: return false;
     }